Finished Draft
compiled all of kevin and yusens work, finished my analysis of results. Cleaned up and added a bit of stuff to the code to make the analysis more sound, also added a bit of visualization to the simulation based section. "Summary Statistic" section isn't done, I need to redo it tomorrow but it should be fast.